requirements.txt
fastapi
uvicorn
openai-whisper
numpy
rapidfuzz
websockets
pydub
sounddevice